Fix compilation on FreeBSDX

+++ utils/storage/storage_unix.go
@@ -14,6 +14,7 @@ func AvailableBytes(storagePath string) (uint64, error) {
 	if err != nil {
 		return 0, err
 	}
-	avail := stat.Bavail * uint64(stat.Bsize)
+	//	avail := stat.Bavail * uint64(stat.Bsize)
+	avail := uint64(stat.Bavail * int64(stat.Bsize))
 	return avail, nil
 }
